requirement.txt
toml >= 0.10.0
tqdm>=4.43.0
pyyaml>=5.3.1
scikit-learn>=0.20.2
xgboost==0.81
pandas>=0.24.1
numpy>=1.16.1
dill>=0.2.9
pytest-cov>=2.5, <3.0
pytest-mock>=1.7.1,<2.0
pytest>=3.4, <4.0
flake8>=3.7.9
black==19.10b0
isort>=4.3.21
autopep8>=1.5.1